test(sessions): keep global compact e2e inside the existing gateway shard
A new gateway test root pushed the core tsgo graph over the 720-file cap.
